+++ cotton-ssdp/examples/ssdp-search-mio.rs
@@ -23,7 +23,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         uuid.to_string(),
         cotton_ssdp::Advertisement {
             notification_type: "test".to_string(),
-            location: url::Url::parse("http://127.0.0.1/test").unwrap(),
+            location: "http://127.0.0.1/test".to_string(),
         },
     );