+++ cotton-ssdp/tests/service.rs
@@ -19,8 +19,7 @@ fn services_can_communicate() {
         "uuid:999",
         Advertisement {
             notification_type: "upnp::Directory:3".to_string(),
-            location: url::Url::parse("http://127.0.0.1/description.xml")
-                .unwrap(),
+            location: "http://127.0.0.1/description.xml".to_string(),
         },
     );
 
@@ -77,8 +76,7 @@ fn services_can_communicate_unicast() {
         "uuid:999",
         Advertisement {
             notification_type: "upnp::Directory:3".to_string(),
-            location: url::Url::parse("http://127.0.0.1/description.xml")
-                .unwrap(),
+            location: "http://127.0.0.1/description.xml".to_string(),
         },
     );
 
@@ -158,8 +156,7 @@ fn services_can_deadvertise() {
         "uuid:998",
         Advertisement {
             notification_type: "upnp::Directory:4".to_string(),
-            location: url::Url::parse("http://127.0.0.1/description.xml")
-                .unwrap(),
+            location: "http://127.0.0.1/description.xml".to_string(),
         },
     );